\name{CIPlib-package}
\alias{CIPlib-package}
\alias{CIPlib}
\docType{package}
\title{
\packageTitle{CIPlib}
}
\description{
\packageDescription{CIPlib}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CIPlib}
\packageIndices{CIPlib}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{CIPlib}

Maintainer: \packageMaintainer{CIPlib}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
